pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "autoref-py"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "bancho.py>=0.1.4",   # IRC connection to Bancho
    "pandas",      # match-state DataFrame
    "aiosu",       # osu! API v2 — mods/gamemode types used in core models
]

[project.optional-dependencies]
# pip install -e ".[web]"   → adds the web UI
web = [
    "fastapi",
    "uvicorn[standard]",
]
# pip install -e ".[plots]" → matplotlib-based stat graphs
plots = [
    "matplotlib",
    "scipy",
]
# pip install -e ".[pp]"    → local pp calculator (rosu-pp-py, Rust)
pp = [
    "rosu-pp-py>=4.0",
]
# pip install -e ".[dev]"   → adds test + env-loading tooling
dev = [
    "pytest",
    "pytest-asyncio",
    "python-dotenv",
]
# pip install -e ".[all]"   → everything
all = [
    "autoref-py[web,plots,pp,dev]",
]

[tool.setuptools.packages.find]
include = ["autoref*"]
